app/requirements_slim.txt
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dotenv==0.9.9
Flask==3.1.2
grpcio==1.76.0
grpcio-health-checking==1.76.0
grpcio-tools==1.75.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
#opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
protobuf==6.32.1
python-dotenv==1.2.1
redis==6.4.0
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
